+++ packages/create-lila-app/src/create.js
@@ -4,6 +4,16 @@ import cp from 'child_process';
 import shell from 'shelljs';
 import fse from 'fs-extra';
 import { error, info, log } from '../../../util/logger';
+import {
+  normalType,
+  reactType,
+  vueType,
+  reactVueType,
+  normalLibType,
+  reactLibType,
+  vueLibType,
+  rollupType,
+} from './data';
 
 const { join } = path;
 const { existsSync } = fs;
@@ -49,7 +59,12 @@ export default ({ dir, type }) => {
 
     content = content.replace('[project-name]', dir);
 
-    if (type === 'webpack-lib' || type === 'rollup') {
+    if (
+      type === normalLibType ||
+      type === reactLibType ||
+      type === vueLibType ||
+      type === rollupType
+    ) {
       content = content.replace(
         '"placeholder": "placeholder",',
         `"main": "build/cjs.js",
@@ -116,10 +131,20 @@ export default ({ dir, type }) => {
 
     const lilaPkg = ['lila-bin', 'lila-core', 'lila-tasks'];
 
-    if (type === 'webpack') lilaPkg.push('lila-webpack', 'lila-webpack-config');
-    else if (type === 'webpack-lib')
+    if (
+      type === normalType ||
+      type === reactType ||
+      type === vueType ||
+      type === reactVueType
+    )
+      lilaPkg.push('lila-webpack', 'lila-webpack-config');
+    else if (
+      type === normalLibType ||
+      type === reactLibType ||
+      type === vueLibType
+    )
       lilaPkg.push('lila-webpack-lib', 'lila-webpack-lib-config');
-    else if (type === 'rollup')
+    else if (type === rollupType)
       lilaPkg.push('lila-rollup', 'lila-rollup-config');
 
     const child2 = spawn(npm, ['install', '--save-dev', ...lilaPkg], {